refactor : 의존관계 정리 및 패키지 구조 변경. 서비스가 심판과 점수판을 의존하고 심판이 점수판을 의존하기 때문에 심판과 점수판과의 의존관계를 삭제함
