+++ src/main/java/baseball/Application.java
@@ -1,11 +1,11 @@
 package baseball;
 
-import baseball.domain.ball.BallValidator;
-import baseball.domain.object.Player;
-import baseball.domain.object.PlayerImpl;
-import baseball.domain.object.Referee;
-import baseball.domain.object.RefereeImpl;
-import baseball.domain.service.GameService;
+import baseball.domain.BallValidator;
+import baseball.domain.Player;
+import baseball.domain.PlayerImpl;
+import baseball.domain.Referee;
+import baseball.domain.RefereeImpl;
+import baseball.service.GameService;
 import baseball.game.GameController;
 import baseball.game.ballgen.InputBallGenerator;
 import baseball.game.ballgen.RandomBallGenerator;
@@ -16,7 +16,7 @@ public class Application {
     public static void main(String[] args) {
         Player hitter = new PlayerImpl(new InputBallGenerator());
         Player pitcher = new PlayerImpl(new RandomBallGenerator(BALL_SIZE, BALL_PIECE_MIN_NUMBER, BALL_PIECE_MAX_NUMBER));
-        Referee referee = new RefereeImpl(BALL_SIZE);
+        Referee referee = new RefereeImpl();
 
         GameService gameService = new GameService(pitcher, hitter, referee);
         gameService.setBallValidator(new BallValidator(BALL_SIZE, BALL_PIECE_MIN_NUMBER, BALL_PIECE_MAX_NUMBER));
