+++ src/main/java/baseball/domain/Player.java
@@ -1,0 +1,5 @@
+package baseball.domain;
+
+public interface Player {
+    Ball generateBall();
+}
