+++ src/main/java/baseball/domain/Referee.java
@@ -1,0 +1,7 @@
+package baseball.domain;
+
+import java.util.List;
+
+public interface Referee {
+    List<String> calcScore(Ball pitchBall, Ball swingBall);
+}
