--- src/main/java/baseball/domain/object/Player.java
@@ -1,7 +1,0 @@
-package baseball.domain.object;
-
-import baseball.domain.ball.Ball;
-
-public interface Player {
-    Ball generateBall();
-}
