--- src/main/java/baseball/domain/object/Referee.java
@@ -1,8 +1,0 @@
-package baseball.domain.object;
-
-import baseball.domain.ball.Ball;
-import baseball.domain.score.Score;
-
-public interface Referee {
-    Score calcScore(Ball pitchBall, Ball swingBall);
-}
