+++ src/main/java/baseball/game/ballgen/InputBallGenerator.java
@@ -1,12 +1,12 @@
 package baseball.game.ballgen;
 
-import baseball.domain.ball.Ball;
-import baseball.domain.ball.BallGenerator;
-import camp.nextstep.edu.missionutils.Console;
+import baseball.domain.Ball;
+import baseball.domain.BallGenerator;
+import baseball.game.utils.UiUtils;
 
 public class InputBallGenerator implements BallGenerator {
     @Override
     public Ball generate() {
-        return new Ball(Console.readLine());
+        return new Ball(UiUtils.readLine());
     }
 }
